Updated packages structure
