+++ src/main/java/kapablankaNew/JeuroNet/Recurrent/GRU/GruLayer.java
@@ -1,10 +1,11 @@
-package kapablankaNew.JeuroNet.Recurrent;
+package kapablankaNew.JeuroNet.Recurrent.GRU;
 
 import kapablankaNew.JeuroNet.Mathematical.ActivationFunction;
 import kapablankaNew.JeuroNet.Mathematical.Matrix;
 import kapablankaNew.JeuroNet.Mathematical.Vector;
 import kapablankaNew.JeuroNet.Mathematical.VectorMatrixException;
 import kapablankaNew.JeuroNet.Mathematical.VectorType;
+import kapablankaNew.JeuroNet.Recurrent.Interfaces.AbstractRecurrentLayer;
 import lombok.EqualsAndHashCode;
 import lombok.NonNull;
