+++ src/main/java/kapablankaNew/JeuroNet/Recurrent/GRU/GruLayerTopology.java
@@ -1,6 +1,9 @@
-package kapablankaNew.JeuroNet.Recurrent;
+package kapablankaNew.JeuroNet.Recurrent.GRU;
 
 import kapablankaNew.JeuroNet.Mathematical.VectorMatrixException;
+import kapablankaNew.JeuroNet.Recurrent.Interfaces.AbstractRecurrentLayerTopology;
+import kapablankaNew.JeuroNet.Recurrent.Interfaces.RecurrentLayer;
+import kapablankaNew.JeuroNet.Recurrent.RecurrentLayerType;
 import kapablankaNew.JeuroNet.TopologyException;
 import lombok.Builder;
 import lombok.EqualsAndHashCode;
@@ -9,12 +12,12 @@ import lombok.Getter;
 import java.io.Serializable;
 
 @EqualsAndHashCode(callSuper = true)
-public class LstmLayerTopology extends AbstractRecurrentLayerTopology implements Serializable {
+public class GruLayerTopology extends AbstractRecurrentLayerTopology implements Serializable {
     @Getter
     private final int hiddenSize;
 
     @Builder
-    private LstmLayerTopology(int inputSize, int outputCount, int outputSize, int hiddenSize,
+    private GruLayerTopology(int inputSize, int outputCount, int outputSize, int hiddenSize,
                              double learningRate, RecurrentLayerType recurrentLayerType) throws TopologyException {
         super(inputSize, outputCount, outputSize, learningRate, recurrentLayerType);
         this.hiddenSize = hiddenSize;
@@ -22,7 +25,6 @@ public class LstmLayerTopology extends AbstractRecurrentLayerTopology implements
 
     @Override
     public RecurrentLayer createRecurrentLayer() throws VectorMatrixException {
-        return new LstmLayer(this);
+        return new GruLayer(this);
     }
 }
-
