+++ src/main/java/kapablankaNew/JeuroNet/Recurrent/LSTM/LstmLayerTopology.java
@@ -1,6 +1,9 @@
-package kapablankaNew.JeuroNet.Recurrent;
+package kapablankaNew.JeuroNet.Recurrent.LSTM;
 
 import kapablankaNew.JeuroNet.Mathematical.VectorMatrixException;
+import kapablankaNew.JeuroNet.Recurrent.Interfaces.AbstractRecurrentLayerTopology;
+import kapablankaNew.JeuroNet.Recurrent.Interfaces.RecurrentLayer;
+import kapablankaNew.JeuroNet.Recurrent.RecurrentLayerType;
 import kapablankaNew.JeuroNet.TopologyException;
 import lombok.Builder;
 import lombok.EqualsAndHashCode;
@@ -9,12 +12,12 @@ import lombok.Getter;
 import java.io.Serializable;
 
 @EqualsAndHashCode(callSuper = true)
-public class GruLayerTopology extends AbstractRecurrentLayerTopology implements Serializable {
+public class LstmLayerTopology extends AbstractRecurrentLayerTopology implements Serializable {
     @Getter
     private final int hiddenSize;
 
     @Builder
-    private GruLayerTopology(int inputSize, int outputCount, int outputSize, int hiddenSize,
+    private LstmLayerTopology(int inputSize, int outputCount, int outputSize, int hiddenSize,
                              double learningRate, RecurrentLayerType recurrentLayerType) throws TopologyException {
         super(inputSize, outputCount, outputSize, learningRate, recurrentLayerType);
         this.hiddenSize = hiddenSize;
@@ -22,6 +25,7 @@ public class GruLayerTopology extends AbstractRecurrentLayerTopology implements 
 
     @Override
     public RecurrentLayer createRecurrentLayer() throws VectorMatrixException {
-        return new GruLayer(this);
+        return new LstmLayer(this);
     }
 }
+
