+++ src/test/java/kapablankaNew/JeuroNet/Recurrent/RecurrentNetworkTest.java
@@ -9,6 +9,10 @@ import kapablankaNew.JeuroNet.Mathematical.LossFunction;
 import kapablankaNew.JeuroNet.Mathematical.Vector;
 import kapablankaNew.JeuroNet.Mathematical.VectorMatrixException;
 import kapablankaNew.JeuroNet.Mathematical.VectorType;
+import kapablankaNew.JeuroNet.Recurrent.GRU.GruLayerTopology;
+import kapablankaNew.JeuroNet.Recurrent.Interfaces.RecurrentLayerTopology;
+import kapablankaNew.JeuroNet.Recurrent.LSTM.LstmLayerTopology;
+import kapablankaNew.JeuroNet.Recurrent.RNN.RnnLayerTopology;
 import kapablankaNew.JeuroNet.TextConverter;
 import kapablankaNew.JeuroNet.TopologyException;
 import org.junit.jupiter.api.BeforeAll;
